refactor: extract project-sessions.js from project.js (PR-11)
Move session management, model/config settings, update management,
process control, rewind/fork, permission response, project management,
and daemon config handlers into a dedicated module. project.js reduced
from 2,878 to 1,857 lines.
